+++ examples/faker/src/gen/mocks/storeMocks/createGetOrderById.ts
@@ -10,7 +10,7 @@ import { GetOrderByIdQueryResponse } from '../../models/GetOrderById'
  * @description Invalid ID supplied
  */
 
-export function createGetOrderById400(): GetOrderById400 {
+export function createGetOrderById400(): NonNullable<GetOrderById400> {
   return undefined
 }
 
@@ -18,11 +18,11 @@ export function createGetOrderById400(): GetOrderById400 {
  * @description Order not found
  */
 
-export function createGetOrderById404(): GetOrderById404 {
+export function createGetOrderById404(): NonNullable<GetOrderById404> {
   return undefined
 }
 
-export function createGetOrderByIdPathParams(): GetOrderByIdPathParams {
+export function createGetOrderByIdPathParams(): NonNullable<GetOrderByIdPathParams> {
   return { orderId: faker.number.float({}) }
 }
 
@@ -30,6 +30,6 @@ export function createGetOrderByIdPathParams(): GetOrderByIdPathParams {
  * @description successful operation
  */
 
-export function createGetOrderByIdQueryResponse(): GetOrderByIdQueryResponse {
+export function createGetOrderByIdQueryResponse(): NonNullable<GetOrderByIdQueryResponse> {
   return createOrder()
 }