+++ packages/swagger-ts/src/generators/TypeGenerator.ts
@@ -82,7 +82,7 @@ export class TypeGenerator extends SchemaGenerator<Options, OpenAPIV3.SchemaObje
     const node = createTypeAliasDeclaration({
       modifiers: [modifiers.export],
       name: this.options.resolveName({ name: baseName, pluginName }) || baseName,
-      type: keysToOmit?.length ? createOmitDeclaration({ keys: keysToOmit, type }) : type,
+      type: keysToOmit?.length ? createOmitDeclaration({ keys: keysToOmit, type, nonNullable: true }) : type,
     })
 
     if (description) {
@@ -112,14 +112,14 @@ export class TypeGenerator extends SchemaGenerator<Options, OpenAPIV3.SchemaObje
    * Delegates to getBaseTypeFromSchema internally and
    * optionally adds a union with null.
    */
-  private getTypeFromSchema(schema: OpenAPIV3.SchemaObject, name?: string): ts.TypeNode | null {
+  private getTypeFromSchema(schema?: OpenAPIV3.SchemaObject, name?: string): ts.TypeNode | null {
     const type = this.getBaseTypeFromSchema(schema, name)
 
     if (!type) {
       return null
     }
 
-    if (schema) {
+    if (schema && !schema.nullable) {
       return type
     }