+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [1.11.6](https://github.com/kubb-project/kubb/compare/kubb-v1.11.5...kubb-v1.11.6) (2023-10-16)
+
+
+### ✨ Features
+
+* **swagger-ts:** support nullable types ([5a4f7cf](https://github.com/kubb-project/kubb/commit/5a4f7cf2e9faa063dc2844c98cf738098ce62ebe))
+
+
+### 🐞 Bug Fixes
+
+* nonnullable for generated types ([0c676e4](https://github.com/kubb-project/kubb/commit/0c676e4ec3735e3f41e6e3bed45dc4bbbcc89a00))
+
+
+### 📦 Miscellaneous Chores
+
+* release 1.11.6 ([70826b2](https://github.com/kubb-project/kubb/commit/70826b2d52d970b16a7f00a8c63f95354699df7c))
+
 ## [1.11.5](https://github.com/kubb-project/kubb/compare/kubb-v1.11.4...kubb-v1.11.5) (2023-10-16)