+++ packages/swagger-faker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.11.6](https://github.com/kubb-project/kubb/compare/@kubb/swagger-faker-v1.11.5...@kubb/swagger-faker-v1.11.6) (2023-10-16)
+
+
+### 🐞 Bug Fixes
+
+* nonnullable for generated types ([0c676e4](https://github.com/kubb-project/kubb/commit/0c676e4ec3735e3f41e6e3bed45dc4bbbcc89a00))
+
+
+### 📦 Miscellaneous Chores
+
+* release 1.11.6 ([70826b2](https://github.com/kubb-project/kubb/commit/70826b2d52d970b16a7f00a8c63f95354699df7c))
+
 ## [1.11.5](https://github.com/kubb-project/kubb/compare/@kubb/swagger-faker-v1.11.4...@kubb/swagger-faker-v1.11.5) (2023-10-16)